Change streamer() to return a tuple instead of only the receiver.
Returning a JoinHandle for the sender in addition to the receiver
permits the calling application to await on the JoinHandle, and deal
with any errors that may have occurred on the sender side.

One motivation for this is to allow the caller to terminate the stream
by dropping the receiver, and allow the sender to exit gracefully.